Small improvements in different layers.

+++ pydtnn/activations/tanh.py
@@ -25,4 +25,4 @@ class Tanh(Activation):
 
     def initialize(self, prev_shape):
         super().initialize(prev_shape)
-        self.out = np.empty(shape=self.shape)
+        self.out = np.empty(shape=self.shape, dtype=self.model.dtype)
